% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmigrowth-dataset.R
\docType{data}
\name{bmigrowth}
\alias{bmigrowth}
\title{BMI Measurements For 100 Individuals From 0 To 17 Years.}
\format{
A data frame with 1050 rows and 6 variables:
\describe{
\item{ID}{(character) ID using three digits.}
\item{age}{(numeric) age in years.}
\item{sex}{(integer) sex with 1: male and 0: female.}
\item{weight}{(numeric) weight in kilograms.}
\item{height}{(integerà height in centimetres.}
\item{bmi}{(numeric) Body Mass Index in kilograms per quare metre.}
}
}
\usage{
bmigrowth
}
\description{
A dataset containing the age, sex, weight, height and BMI for 100 individuals.
Measurements performed from birth to 17 years old.
}
\keyword{datasets}
