% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_apar.R
\name{compute_apar}
\alias{compute_apar}
\title{Compute adiposity peak (AP) and adiposity rebound (AR).}
\usage{
compute_apar(
  fit,
  from = c("predicted", "observed"),
  start = 0.25,
  end = 10,
  step = 0.05
)
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call \code{nlme::lme()}, \code{time_model()} or \code{egg_model()}.}

\item{from}{A string indicating the type of data to be used for the AP and AR
computation, either "predicted" or "observed". Default is "predicted".}

\item{start}{The start of the time window to compute AP and AR.}

\item{end}{The end of the time window to compute AP and AR.}

\item{step}{The step to increment the sequence.}
}
\value{
A \code{data.table} object.
}
\description{
Compute adiposity peak (AP) and adiposity rebound (AR).
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)

head(compute_apar(fit = res, from = "predicted")[AP | AR])
}
