% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_auc.R
\name{compute_auc}
\alias{compute_auc}
\title{compute_auc}
\usage{
compute_auc(
  fit,
  method,
  period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]]
)
}
\arguments{
\item{fit}{something}

\item{method}{something}

\item{period}{something}

\item{knots}{something}
}
\value{
something
}
\description{
compute_auc
}
