% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_slopes.R
\name{compute_slopes}
\alias{compute_slopes}
\title{Predict average slopes for several intervals using a model fitted by \code{time_model}.}
\usage{
compute_slopes(
  fit,
  method,
  period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]]
)
}
\arguments{
\item{fit}{A model object from a statistical model such as from a call to \code{time_model()}.}

\item{method}{The type of model provided in \code{fit},
\emph{i.e.}, one of \code{"cubic_slope"}, \code{"linear_splines"} or \code{"cubic_splines"}.}

\item{period}{The intervals knots on which slopes are to be computed.}

\item{knots}{The knots as defined \code{fit} and according to \code{method}.}
}
\value{
A \code{data.frame} with slopes for each individuals/samples.
}
\description{
Comoute average slopes for "clubic slope", "linear splines" and "cubic splines"
fitted using \code{time_model}.
}
\examples{
data("bmigrowth")
ls_mod <- time_model(
  x = "age",
  y = "log(bmi)",
  cov = NULL,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  method = "linear_splines"
)
head(compute_slopes(
  fit = ls_mod,
  method = "linear_splines",
  period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17)#,
  # knots = list(
  #   "cubic_slope" = NULL,
  #   "linear_splines" = c(5.5, 11),
  #   "cubic_splines" = c(2, 8, 12)
  # )[[method]]
))
}
