% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_eggla_gwas.R
\name{do_eggla_gwas}
\alias{do_eggla_gwas}
\title{Perform GWAS using PLINK2 (and BCFtools)}
\usage{
do_eggla_gwas(
  data,
  results_zip,
  id_column,
  formula,
  vcfs,
  vep = NULL,
  path,
  bin_path = list(bcftools = "/usr/bin/bcftools", plink2 =
    "https://s3.amazonaws.com/plink2-assets/plink2_linux_x86_64_20220410.zip"),
  threads = 1
)
}
\arguments{
\item{data}{Path to the phenotypes stored as a CSV file.}

\item{results_zip}{Path to the zip archives generated by \code{run_eggla()}.}

\item{id_column}{Name of the column where sample/individual IDs are stored.}

\item{formula}{Formula to be used for the analysis,
with on the left hand side, one or multiple traits to be analysed separately,
and on the right hand side, the covariates to be used,
\emph{i.e.}, columns' names from \code{data}.}

\item{vcfs}{Path to the "raw" VCF file(s) containing
the genotypes of the individuals to be analysed.}

\item{vep}{Path to the VEP annotation file to be used to set variants RSIDs and add gene SYMBOL, etc.}

\item{path}{Directory in which computation will occur and where output files will be saved.}

\item{bin_path}{A named list containing the path to the PLINK2 and BCFtools binaries
For PLINK2, an URL to the binary can be provided (see https://www.cog-genomics.org/plink/2.0).}

\item{threads}{Number of threads to be used by BCFtools}
}
\value{
Path to results file.
}
\description{
Format VCF file(s) by filtering out all variants
not satisfaying "--min-alleles 2 --max-alleles 2 --types snps"
and setting IDs (if no annotation file using VEP is provided)
with "\%CHROM:\%POS:\%REF:\%ALT" (see https://samtools.github.io/bcftools/).
GWAS is performed on the formatted VCF file(s) by PLINK2 software
(https://www.cog-genomics.org/plink/2.0).
}
\examples{
if (interactive()) {
  data("bmigrowth")
  fwrite(
    x = bmigrowth,
    file = file.path(tempdir(), "bmigrowth.csv")
  )
  results_archives <- run_eggla(
    data = fread(file.path(tempdir(), "bmigrowth.csv")),
    id_variable = "ID",
    age_days_variable = NULL,
    age_years_variable = "age",
    weight_kilograms_variable = "weight",
    height_centimetres_variable = "height",
    sex_variable = "sex",
    covariates = NULL,
    male_coded_zero = FALSE,
    random_complexity = 1,
    parallel = FALSE,
    parallel_n_chunks = 1,
    working_directory = tempdir()
  )
  do_eggla_gwas(
    data = "/tmp/bmigrowth.csv",
    results_zip = results_archives,
    id_column = "ID",
    formula = `slope_0--0.5` + `slope_1.5--5` + `auc_0--0.5` + `auc_1.5--5` ~ sex,
    vcfs = list.files(
      path = file.path(tempdir(), "vcf"),
      pattern = "\\\\.vcf$|\\\\.vcf.gz$",
      full.names = TRUE
    ),
    vep = NULL,
    path = tempdir(),
    bin_path = list(
      bcftools = "/usr/bin/bcftools",
      plink2 = "/usr/bin/plink2"
    ),
    threads = 1
  )
}
}
