% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_aucs.R
\name{egg_aucs}
\alias{egg_aucs}
\title{Derived areas under the curve from a cubic splines mixed-effects model by \code{time_model}.}
\usage{
egg_aucs(fit, period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17))
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{egg_model()}.}

\item{period}{The intervals knots on which AUCs are to be computed.}
}
\value{
A \code{data.frame} with AUC for each individuals/samples.
}
\description{
Derived areas under the curve (AUCs) for differentintervals based
on a fitted cubic splines mixed-effects model from \code{egg_model}.
This function is a specific version of \code{compute_aucs}
designed to work specifically on \code{egg_model}.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
head(
  egg_aucs(
    fit = res,
    period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17)
  )
)
}
