% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_correlations.R
\name{egg_correlations}
\alias{egg_correlations}
\title{Compute the derived parameters correlations from a cubic splines mixed-effects model by \code{egg_model()}.}
\usage{
egg_correlations(
  fit,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  knots = c(1, 8, 12)
)
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{egg_model()}.}

\item{period}{The intervals knots on which slopes are to be computed.}

\item{knots}{The knots as defined \code{fit} and according to \code{method}.}
}
\value{
A \code{ggplot2} object with correlations between each intervals derived parameters.
}
\description{
Based on computed area under the curves (\emph{i.e.}, \code{egg_aucs()})
and slopes (\emph{i.e.}, \code{egg_slopes()}) for several intervals using
a model fitted by \code{egg_model()}, compute the correlations between
each intervals derived parameters.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
egg_correlations(
  fit = res,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  knots = c(1, 8, 12)
)
}
