% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_model.R
\name{egg_model}
\alias{egg_model}
\title{egg_model}
\usage{
egg_model(formula, data)
}
\arguments{
\item{data}{A data.frame containing the variables named in \code{x} and \code{y}.}

\item{x}{A length one character vector with the main covariate name (\emph{i.e.}, right-hand side).}

\item{y}{A length one character vector with the variable name to be explained (\emph{i.e.}, left-hand side).}

\item{cov}{A vector of addtional/optional covariates names to included in the fixed effect part of the linear mixed-effects models.}
}
\value{
An object of class "lme" representing the linear mixed-effects model fit.
}
\description{
egg_model
}
