% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_slopes.R
\name{egg_slopes}
\alias{egg_slopes}
\title{Derived slopes from a cubic splines mixed-effects model by \code{time_model()}.}
\usage{
egg_slopes(fit, period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17))
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{egg_model()}.}

\item{period}{The intervals knots on which slopes are to be computed.}
}
\value{
A \code{data.frame} with slopes for each individuals/samples.
}
\description{
Derived slopes for different intervals based on a fitted
cubic splines mixed-effects model from \code{egg_model()}.
This function a specific version of \code{compute_slopes}
designed to work specifically on \code{egg_model()}.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
head(
  egg_slopes(
    fit = res,
    period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17)
  )
)
}
