% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggla-package.R
\docType{package}
\name{eggla-package}
\alias{eggla}
\alias{eggla-package}
\title{eggla: Early Growth Genetics Longitudinal Analysis}
\description{
Tools for longitudinal analysis within the EGG (Early Growth Genetics) Consortium.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/eggla}
  \item Report bugs at \url{https://github.com/mcanouil/eggla/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{***REMOVED***} (\href{https://orcid.org/0000-0002-3396-4549}{ORCID})

Authors:
\itemize{
  \item Nicole Warrington \email{n.warrington@uq.edu.au}
}

}
\keyword{internal}
