% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_egg_aucs.R
\name{plot_egg_aucs}
\alias{plot_egg_aucs}
\title{Plot derived area under the curves from a model fitted by \code{egg_model()}.}
\usage{
plot_egg_aucs(
  fit,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  knots = c(1, 8, 12)
)
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{egg_model()}.}

\item{period}{The intervals knots on which AUCs are to be computed.}

\item{knots}{The knots as defined \code{fit} and according to \code{method}.}
}
\value{
A \code{patchwork} \code{ggplot2} object.
}
\description{
Plot derived area under the curves from a model fitted by \code{egg_model()}.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
plot_egg_aucs(
  fit = res,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17)
) +
  theme_minimal() +
  theme(
    axis.line = element_line(colour = "grey20", size = rel(1)),
    axis.ticks = element_line(colour = "grey20"),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.x = element_blank()
  )
}
