% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_egg_slopes.R
\name{plot_egg_slopes}
\alias{plot_egg_slopes}
\title{Plot derived slopes from a model fitted by \code{egg_model()}.}
\usage{
plot_egg_slopes(fit, period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17))
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{egg_model()}.}

\item{period}{The intervals knots on which AUCs are to be computed.}
}
\value{
A \code{patchwork} \code{ggplot2} object.
}
\description{
Plot derived slopes from a model fitted by \code{egg_model()}.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
plot_egg_slopes(
  fit = res,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17)
)
}
