% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slopes.R
\name{plot_slopes}
\alias{plot_slopes}
\title{Plot derived slopes from a model fitted by \code{time_model()}.}
\usage{
plot_slopes(
  fit,
  method,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(1, 8,
    12))[[method]]
)
}
\arguments{
\item{fit}{A model object from a statistical model such as
from a call to \code{time_model()}.}

\item{method}{The type of model provided in \code{fit},
\emph{i.e.}, one of \code{"cubic_slope"}, \code{"linear_splines"} or \code{"cubic_splines"}.}

\item{period}{The intervals knots on which AUCs are to be computed.}

\item{knots}{The knots as defined \code{fit} and according to \code{method}.}
}
\value{
A \code{patchwork} \code{ggplot2} object.
}
\description{
Plot derived slopes from a model fitted by \code{time_model()}.
}
\examples{
library(ggplot2)
data("bmigrowth")
ls_mod <- time_model(
  x = "age",
  y = "log(bmi)",
  cov = NULL,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  method = "linear_splines"
)
plot_slopes(
  fit = ls_mod,
  method = "linear_splines"
) +
  theme_minimal() +
  theme(
    axis.line = element_line(colour = "grey20", size = rel(1)),
    axis.ticks = element_line(colour = "grey20"),
    panel.grid.minor.x = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.x = element_blank()
  )
}
