% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_average_slopes.R
\name{predict_average_slopes}
\alias{predict_average_slopes}
\title{predict_average_slopes}
\usage{
predict_average_slopes(
  fit,
  method,
  period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]]
)
}
\arguments{
\item{fit}{something}

\item{method}{something}

\item{period}{something}

\item{knots}{something}
}
\value{
something
}
\description{
predict_average_slopes
}
