% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eggla.R
\name{run_eggla}
\alias{run_eggla}
\title{run_eggla}
\usage{
run_eggla(
  phenotypes = NULL,
  trait = "bmi",
  covariates = NULL,
  daymont_qc = TRUE,
  daymont_flags_exclude = TRUE,
  cohort_name = "cohort",
  output_directory = tempdir(),
  ...
)
}
\arguments{
\item{phenotypes}{something}

\item{trait}{something}

\item{covariates}{something}

\item{daymont_qc}{something}

\item{daymont_flags_exclude}{something}

\item{cohort_name}{something}

\item{output_directory}{something}

\item{...}{Parameters to pass to \code{rmarkdown::render()}.}
}
\description{
run_eggla
}
\keyword{internal}
