% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eggla_lmm.R
\name{run_eggla_lmm}
\alias{run_eggla_lmm}
\title{Perform EGG longitudinal analysis and derived areas under the curves and slopes.}
\usage{
run_eggla_lmm(
  data,
  id_variable,
  age_days_variable,
  age_years_variable,
  weight_kilograms_variable,
  height_centimetres_variable,
  sex_variable,
  covariates,
  male_coded_zero = FALSE,
  random_complexity = "auto",
  use_car1 = FALSE,
  knots = c(1, 8, 12),
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  filter_apar = NULL,
  outlier_method = "iqr",
  outlier_threshold = list(iqr = 1.7),
  outlier_exclude = TRUE,
  parallel = FALSE,
  parallel_n_chunks = 1,
  working_directory = getwd(),
  quiet = FALSE,
  clean = TRUE
)
}
\arguments{
\item{data}{Phenotypes data that inherits from \code{data.frame} class.}

\item{id_variable}{Name of the column where sample/individual IDs are stored.}

\item{age_days_variable}{Name of the column where age in days is stored.
\code{NULL} if age in days is not available.}

\item{age_years_variable}{Name of the column where age in years is stored.
\code{NULL} if age in years is not available.}

\item{weight_kilograms_variable}{Name of the column where weight in kilograms is stored.}

\item{height_centimetres_variable}{Name of the column where height in centimetres is stored.}

\item{sex_variable}{Name of the column where sex is stored.}

\item{covariates}{A vector of columns' names to be used as covariates.
\code{NULL} if there are no covariates to add.}

\item{male_coded_zero}{Is male coded "0" (and female coded "1")?}

\item{random_complexity}{A numeric (1-3) indicating the complexity of the random effect term.
Default, \code{"auto"} will try from the more complex to the less complex if no success.}

\item{use_car1}{A logical indicating whether to use continuous auto-correlation,
i.e., CAR(1) as correlation structure.}

\item{knots}{The knots defining the splines.}

\item{period}{The intervals knots on which slopes are to be computed.}

\item{filter_apar}{A string following \code{data.table} syntax for filtering on \code{"i"}
(\emph{i.e.}, row elements), \emph{e.g.}, \code{filter = "source == 'A'"}.
Argument pass through \code{compute_apar()} (see \code{predict_bmi()}).
Default is \code{NULL}.}

\item{outlier_method}{The outlier detection method(s). Can be \code{"all"} or some of
\code{"cook"}, \code{"pareto"}, \code{"zscore"}, \code{"zscore_robust"}, \code{"iqr"}, \code{"ci"}, \code{"eti"},
\code{"hdi"}, \code{"bci"}, \code{"mahalanobis"}, \code{"mahalanobis_robust"}, \code{"mcd"}, \code{"ics"},
\code{"optics"} or \code{"lof"}.
See \code{performance::check_outliers()} \url{https://easystats.github.io/performance/reference/check_outliers.html} for details.}

\item{outlier_threshold}{A list containing the threshold values for each method (\emph{e.g.},
\code{list('mahalanobis' = 7, 'cook' = 1)}), above which an observation is
considered as outlier. If \code{NULL}, default values will be used (see
'Details'). If a numeric value is given, it will be used as the threshold
for any of the method run.
See \code{performance::check_outliers()} \url{https://easystats.github.io/performance/reference/check_outliers.html} for details.}

\item{outlier_exclude}{Whether or not the values/individuals flagged as being outliers should be excluded.
Default is \code{TRUE}.}

\item{parallel}{Determines if \code{growthcleanr::cleangrowth()} function shoud be run in parallel. Defaults to \code{FALSE}.}

\item{parallel_n_chunks}{Specify the number of batches (in \code{growthcleanr::cleangrowth()}) to run in parallel.
Only applies if parallel is set to TRUE.
Defaults to the number of workers returned by the getDoParWorkers function in the foreach package.}

\item{working_directory}{Directory in which computation will occur and where output files will be saved.}

\item{quiet}{A logical indicating whether to suppress the output.}

\item{clean}{A logical indicating whether to clean \code{working_directory} once the archives are created.}
}
\value{
Path to zip archives.
}
\description{
Perform Daymont's quality-control for BMI,
fit a cubic splines mixed model regression
with linear splines as random effect,
save model object, generates residuals figures fot model validity,
derived area under the curve and slopes for male and femal.
This function is a wrapper around \code{egg_model()}, \code{egg_slopes()} and \code{egg_aucs()}.
}
\examples{
if (interactive()) {
  data("bmigrowth")
  fwrite(
    x = bmigrowth,
    file = file.path(tempdir(), "bmigrowth.csv")
  )
  res <- run_eggla_lmm(
    data = fread(file.path(tempdir(), "bmigrowth.csv")),
    id_variable = "ID",
    age_days_variable = NULL,
    age_years_variable = "age",
    weight_kilograms_variable = "weight",
    height_centimetres_variable = "height",
    sex_variable = "sex",
    covariates = NULL,
    random_complexity = 1,
    working_directory = tempdir()
  )
}
}
