% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_to_character.R
\name{sex_to_character}
\alias{sex_to_character}
\title{sex_to_character}
\usage{
sex_to_character(x)
}
\arguments{
\item{x}{A vector of numeric to be converted to character.}
}
\value{
A vector of character.
}
\description{
sex_to_character
}
