% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_model.R
\name{time_model}
\alias{time_model}
\title{time_model}
\usage{
time_model(
  x,
  y,
  cov = NULL,
  data,
  method = c("cubic_slope", "linear_splines", "cubic_splines"),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]]
)
}
\arguments{
\item{x}{something}

\item{y}{something}

\item{cov}{something}

\item{data}{something}

\item{method}{something}

\item{knots}{something}
}
\value{
something
}
\description{
time_model
}
