% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_apar.R
\name{compute_apar}
\alias{compute_apar}
\title{Compute adiposity peak (AP) and adiposity rebound (AR).}
\usage{
compute_apar(fit)
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call \code{nlme::lme()}, \code{time_model()} or \code{egg_model}.}
}
\value{
A \code{data.table} object.
}
\description{
Compute adiposity peak (AP) and adiposity rebound (AR).
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)

head(compute_apar(fit = res)[AP | AR])
}
