% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_apar.R
\name{compute_apar}
\alias{compute_apar}
\title{Compute adiposity peak (AP) and adiposity rebound (AR).}
\usage{
compute_apar(fit, id_var, age_var)
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call \code{nlme::lme()} and \code{time_model()}.}

\item{id_var}{A length one character vector
with the variable name definiing individuals.}

\item{age_var}{A length one character vector
with the variable name definiing the time variable, \emph{i.e.}, age).}
}
\value{
A \code{data.table} object.
}
\description{
Compute adiposity peak (AP) and adiposity rebound (AR).
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
res_apar <- compute_apar(
  fit = res,
  id_var = "ID",
  age_var = "age"
)

head(res_apar[AP | AR])
}
