% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_outliers.R
\name{compute_outliers}
\alias{compute_outliers}
\title{Compute outliers detection in AUCs/Slopes derived parameters.}
\usage{
compute_outliers(data)
}
\arguments{
\item{dt}{A \code{data.frame} with AUC or slopes for each individuals/samples.}
}
\value{
A \code{data.frame} with the individuals and measures flagged as outliers.
}
\description{
Based on computed area under the curves (\emph{i.e.}, \code{compute_aucs()})
and slopes (\emph{i.e.}, \code{compute_slopes()}) for several intervals using
a model fitted by \code{time_model()}, compute an outlier detection for
each intervals derived parameters.
}
\examples{
data("bmigrowth")
ls_mod <- time_model(
  x = "age",
  y = "log(bmi)",
  cov = NULL,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  method = "linear_splines"
)
aucs <- compute_aucs(
  fit = ls_mod,
  method = "linear_splines",
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17)#,
  # knots = list(
  #   "cubic_slope" = NULL,
  #   "linear_splines" = c(5.5, 11),
  #   "cubic_splines" = c(2, 8, 12)
  # )[[method]]
)
compute_outliers(aucs)
}
