% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_model.R
\name{egg_auc}
\alias{egg_auc}
\title{Derived areas under the curve from a cubic splines mixed-effects model.}
\usage{
egg_auc(fit, period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17), id_var = "ID")
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{nlme::lme()} and \code{time_model()}.}

\item{period}{The intervals knots on which AUCs are to be computed.}

\item{id_var}{A string indicating the name of the variable
to be used as the individual identifier.}
}
\value{
A \code{data.frame} with AUC for each individuals/samples.
}
\description{
Derived areas under the curve (AUCs) for differentintervals based
on a fitted cubic splines mixed-effects model from \code{egg_model}.
This function is a specific version of \code{compute_auc}
designed to work specifically on \code{egg_model}.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
head(
  egg_auc(
    fit = res,
    period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17),
    id_var = "ID"
  )
)
}
