% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_model.R
\name{egg_model}
\alias{egg_model}
\title{egg_model}
\usage{
egg_model(formula, data, id_var)
}
\arguments{
\item{formula}{An object of class "formula":
a symbolic description of the model to be fitted with,
time component as the first term in the right-hand side.}

\item{data}{A data.frame containing the variables defined in formula.}

\item{id_var}{A string indicating the name of the variable
to be used as the individual identifier.}
}
\value{
An object of class "lme" representing
the linear mixed-effects model fit.
}
\description{
Fit a cubic splines mixed model regression
with linear splines as random effect.
}
