% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_model.R
\name{egg_model}
\alias{egg_model}
\title{Fit a cubic splines mixed model.}
\usage{
egg_model(
  formula,
  data,
  id_var,
  random_complexity = "auto",
  use_ar1 = FALSE,
  knots = c(2, 8, 12)
)
}
\arguments{
\item{formula}{An object of class "formula":
a symbolic description of the model to be fitted with,
time component as the first term in the right-hand side.}

\item{data}{A data.frame containing the variables defined in formula.}

\item{id_var}{A string indicating the name of the variable
to be used as the individual identifier.}

\item{random_complexity}{A numeric (1-3) indicating the complexity of the random effect term.
Default, \code{"auto"} will try from the more complex to the less complex if no success.}

\item{use_ar1}{A logical indicating whether to use AR(1) for autocorrelation.}

\item{knots}{The knots defining the splines.}
}
\value{
An object of class "lme" representing
the linear mixed-effects model fit.
}
\description{
Fit a cubic splines mixed model regression
with three splines parametrisation as random effect.
This function a specific version of \code{time_model()}.
}
\examples{
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
sres <- as.data.frame(summary(res)[["tTable"]])
rownames(sres) <- sub("gsp\\\\(.*\\\\)\\\\)", "gsp(...)", rownames(sres))
sres
}
