% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_model.R
\name{egg_slopes}
\alias{egg_slopes}
\title{egg_slopes}
\usage{
egg_slopes(fit, period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17), id_var = "ID")
}
\arguments{
\item{fit}{A model object from a statistical model
such as from a call to \code{nlme::lme()} and \code{time_model()}.}

\item{period}{The intervals knots on which slopes are to be computed.}

\item{id_var}{A string indicating the name of the variable
to be used as the individual identifier.}
}
\value{
A \code{data.frame} with slopes for each individuals/samples.
}
\description{
egg_slopes
}
