% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggla-package.R
\docType{package}
\name{eggla-package}
\alias{eggla}
\alias{eggla-package}
\title{eggla: Early Growth Genetics Longitudinal Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for longitudinal analysis within the EGG (Early Growth Genetics) Consortium (\url{http://egg-consortium.org/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/eggla}
  \item \url{https://m.canouil.fr/eggla/}
  \item Report bugs at \url{https://github.com/mcanouil/eggla/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{mickael.canouil@cnrs.fr} (\href{https://orcid.org/0000-0002-3396-4549}{ORCID})

Authors:
\itemize{
  \item Nicole Warrington \email{n.warrington@uq.edu.au} (\href{https://orcid.org/0000-0003-4195-775X}{ORCID})
  \item Kimberley Burrows \email{Kimberley.Burrows@bristol.ac.uk} (\href{https://orcid.org/0000-0002-3208-0389}{ORCID})
  \item Anni Heiskala \email{Anni.Heiskala@oulu.fi}
}

}
\keyword{internal}
