% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(x, y, fit, variables_unit)
}
\arguments{
\item{x}{something}

\item{y}{something}

\item{fit}{something}

\item{variables_unit}{something}
}
\value{
something
}
\description{
plot_residuals
}
