% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot several residuals plots for diagnostics.}
\usage{
plot_residuals(x, y, fit)
}
\arguments{
\item{x}{A length one character vector with the main covariate name
(\emph{i.e.}, right-hand side), as defined in \code{fit}.}

\item{y}{A length one character vector with the variable name to be explained
(\emph{i.e.}, left-hand side), as defined in \code{fit}.}

\item{fit}{A model object from a statistical model
such as from a call \code{time_model()} or \code{egg_model}.}
}
\value{
A \code{patchwork} \code{ggplot2} object.
}
\description{
Plot several residuals plots for diagnostics.
}
\examples{
library(patchwork)
data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
plot_residuals(
  x = "age",
  y = "log(bmi)",
  fit = res
) +
  plot_annotation(
    title = "Cubic Splines (Random Linear Splines) - BMI - Female",
    tag_levels = "A"
  )
}
