% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_average_slopes.R
\name{predict_average_slopes}
\alias{predict_average_slopes}
\title{predict_average_slopes}
\usage{
predict_average_slopes(
  fit,
  method,
  period = c(0, 0.5, 1.5, 5, 6, 10, 12, 17),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]]
)
}
\arguments{
\item{fit}{A model object from a statistical model such as from a call to \code{nlme::lme()} and \code{time_model()}.}

\item{method}{The type of model provided in \code{fit}, \emph{i.e.}, one of \code{"cubic_slope"}, \code{"linear_splines"} or \code{"cubic_splines"}.}

\item{period}{The intervals knots on which slopes are to be computed.}

\item{knots}{The knots as defined \code{fit} and according to \code{method}.}
}
\value{
A \code{data.frame} with slopes for each individuals/samples.
}
\description{
predict_average_slopes
}
