% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eggla.R
\name{run_eggla}
\alias{run_eggla}
\title{run_eggla}
\usage{
run_eggla(
  data,
  id_variable,
  age_days_variable,
  age_years_variable,
  weight_kilograms_variable,
  height_centimetres_variable,
  sex_variable,
  covariates,
  male_coded_zero = FALSE,
  parallel = FALSE,
  parallel_n_chunks = 1,
  working_directory = getwd()
)
}
\arguments{
\item{data}{something}

\item{id_variable}{something}

\item{age_days_variable}{something}

\item{age_years_variable}{something}

\item{weight_kilograms_variable}{something}

\item{height_centimetres_variable}{something}

\item{sex_variable}{something}

\item{covariates}{something}

\item{male_coded_zero}{something}

\item{parallel}{something}

\item{parallel_n_chunks}{something}

\item{working_directory}{something}
}
\value{
Path to zip archives
}
\description{
Perform Daymont's quality-control for BMI,
fit a cubic splines mixed model regression with linear splines as random effect,
save model object, generates residuals figures fot model validity,
derived area under the curve and slopes for male and femal.
}
