% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eggla.R
\name{run_eggla}
\alias{run_eggla}
\title{Perform EGG longitudinal analysis and derived areas under the curves and slopes.}
\usage{
run_eggla(
  data,
  id_variable,
  age_days_variable,
  age_years_variable,
  weight_kilograms_variable,
  height_centimetres_variable,
  sex_variable,
  covariates,
  male_coded_zero = FALSE,
  random_complexity = "auto",
  parallel = FALSE,
  parallel_n_chunks = 1,
  working_directory = getwd()
)
}
\arguments{
\item{data}{Phenotypes data that inherits from \code{data.frame} class.}

\item{id_variable}{Name of the column where sample/individual IDs are stored.}

\item{age_days_variable}{Name of the column where age in days is stored.}

\item{age_years_variable}{Name of the column where age in years is stored.}

\item{weight_kilograms_variable}{Name of the column where weight in kilograms is stored.}

\item{height_centimetres_variable}{Name of the column where height in centimetres is stored.}

\item{sex_variable}{Name of the column where sex is stored.}

\item{covariates}{A vector of columns' names to be used as covariates.}

\item{male_coded_zero}{Is male coded "0" (and female coded "1")?}

\item{random_complexity}{A numeric (1-3) indicating the complexity of the random effect term.
Default, \code{"auto"} will try from the more complex to the less complex if no success.}

\item{parallel}{Determines if \code{growthcleanr::cleangrowth()} function shoud be run in parallel. Defaults to \code{FALSE}.}

\item{parallel_n_chunks}{Specify the number of batches (in \code{growthcleanr::cleangrowth()}) to run in parallel.
Only applies if parallel is set to TRUE.
Defaults to the number of workers returned by the getDoParWorkers function in the foreach package.}

\item{working_directory}{Directory in which computation will occur and where output files will be saved.}
}
\value{
Path to zip archives.
}
\description{
Perform Daymont's quality-control for BMI,
fit a cubic splines mixed model regression
with linear splines as random effect,
save model object, generates residuals figures fot model validity,
derived area under the curve and slopes for male and femal.
This function is a wrapper around \code{egg_model()}, \code{egg_slopes()} and \code{egg_aucs()}.
}
\examples{
if (interactive()) {
  data("bmigrowth")
  fwrite(
    x = bmigrowth,
    file = file.path(tempdir(), "bmigrowth.csv")
  )
  res <- run_eggla(
    data = fread(file.path(tempdir(), "bmigrowth.csv")),
    id_variable = "ID",
    age_days_variable = NULL,
    age_years_variable = "age",
    weight_kilograms_variable = "weight",
    height_centimetres_variable = "height",
    sex_variable = "sex",
    covariates = NULL,
    male_coded_zero = FALSE,
    random_complexity = 1,
    parallel = FALSE,
    parallel_n_chunks = 1,
    working_directory = tempdir()
  )
}
}
