% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eggla.R
\name{run_eggla}
\alias{run_eggla}
\title{run_eggla}
\usage{
run_eggla(
  phenotypes = NULL,
  trait = "bmi",
  cohort_name = "cohort",
  output_directory = tempdir(),
  ...
)
}
\arguments{
\item{phenotypes}{something}

\item{trait}{something}

\item{cohort_name}{something}

\item{output_directory}{something}

\item{...}{Parameters to pass to \code{rmarkdown::render()}.}
}
\description{
run_eggla
}
\keyword{internal}
