% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_model.R
\name{time_model}
\alias{time_model}
\title{time_model}
\usage{
time_model(
  x,
  y,
  data,
  method = c("cubic_slope", "linear_splines", "cubic_splines"),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]],
  as_text = FALSE
)
}
\arguments{
\item{x}{something}

\item{y}{something}

\item{data}{something}

\item{method}{something}

\item{knots}{something}

\item{as_text}{something}
}
\value{
something
}
\description{
time_model
}
