% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_model.R
\name{time_model}
\alias{time_model}
\title{time_model}
\usage{
time_model(
  x,
  y,
  cov = NULL,
  data,
  method = c("cubic_slope", "linear_splines", "cubic_splines"),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(2, 8,
    12))[[method]]
)
}
\arguments{
\item{x}{A length one character vector with the main covariate name (\emph{i.e.}, right-hand side).}

\item{y}{A length one character vector with the variable name to be explained (\emph{i.e.}, left-hand side).}

\item{cov}{A vector of addtional/optional covariates names to included in the fixed effect part of the linear mixed-effects models.}

\item{data}{A data.frame containing the variables named in \code{x} and \code{y}.}

\item{method}{The type of model, \emph{i.e.}, one of \code{"cubic_slope"}, \code{"linear_splines"} or \code{"cubic_splines"}.}

\item{knots}{The knots defining the splines for \code{"linear_splines"} ànd \code{"cubic_splines"} methods.}
}
\value{
An object of class "lme" representing the linear mixed-effects model fit.
}
\description{
time_model
}
\examples{
data("bmigrowth")
ls_mod <- time_model(
  x = "age",
  y = "log(bmi)",
  cov = NULL,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  method = "linear_splines"
)
sres <- as.data.frame(summary(ls_mod)[["tTable"]])
rownames(sres) <- sub("gsp\\\\(.*\\\\)\\\\)", "gsp(...)", rownames(sres))
sres
}
